ref(onboarding): Split pyramid onboarding docs
