+++ static/app/gettingStartedDocs/python/pyramid/index.tsx
@@ -1,0 +1,25 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {
+  feedbackOnboardingJsLoader,
+  replayOnboardingJsLoader,
+} from 'sentry/gettingStartedDocs/javascript/jsLoader/jsLoader';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/python/python/agentMonitoring';
+import {crashReport} from 'sentry/gettingStartedDocs/python/python/crashReport';
+import {featureFlag} from 'sentry/gettingStartedDocs/python/python/featureFlag';
+import {logs} from 'sentry/gettingStartedDocs/python/python/logs';
+import {profiling} from 'sentry/gettingStartedDocs/python/python/profiling';
+
+import {onboarding} from './onboarding';
+
+const docs: Docs = {
+  onboarding,
+  replayOnboardingJsLoader,
+  crashReportOnboarding: crashReport,
+  featureFlagOnboarding: featureFlag,
+  feedbackOnboardingJsLoader,
+  profilingOnboarding: profiling(),
+  agentMonitoringOnboarding: agentMonitoring,
+  logsOnboarding: logs(),
+};
+
+export default docs;
